requirements-test.txt
-r requirements.txt
black
flake8
flake8-bandit
flake8-bugbear
flake8-comprehensions
flake8-docstrings
flake8-logging-format
flake8-pep3101
flake8-print
flake8-rst
flake8-rst-docstrings
isort
mypy
requests-mock